Fix DKMS builds for amd64 kernels on i386 architecture
Parse the kernel version in the DKMS configuration to get the
architecture and kernel version and use that to form the AFS sysname
on i386 and amd64.  This fixes DKMS builds of the module for amd64
kernels on the i386 architecture, and provides a framework that can
be used for similar cases in the future.